compounded -f and -h into one thing in main()
Inspired by how you handled f and h in one set of code. I got rid of the h_flag but kept the f_flag. Basically if f_flag then we do -f things but if !f_flag we do -h things. Not sure how to streamline the code beyond this point though because fundamentally we use different things to check it